refactor(effect): split paint pipeline and group effect header state
paintWindow's shader branch becomes paintShaderTransitionWindow behind a
context struct with byte-equivalent control flow; snapshot capture and the
quad-deform hook move to paint_capture.cpp; trailing header members fold
into DragActivationState/DaemonGateState/IdCacheState value structs.
